[util] Move Validators to a new module
